Cart Delete 기능 작성

+++ app/Http/Controllers/UserCartCartItemsController.php
@@ -3,16 +3,25 @@
 namespace App\Http\Controllers;
 
 use App\Models\Menu;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Exceptions\EaterySyncException;
 use App\Exceptions\OptionCountException;
 use App\Http\Resources\CartItemResource;
 
-class CartItemsController extends Controller
+class UserCartCartItemsController extends Controller
 {
-    public function store()
+    public function store(User $user)
     {
+        if (Auth::user()->id != $user->id) {
+            return response()->json([
+                'status' => 'unauthorized',
+                'message' => '권환이 없습니다.',
+                'error' => '401'
+            ], 401);
+        }
+
         $this->validate(request(), [
             'menu_id' => ['required'],
             'quantity' => ['required', 'numeric'],
@@ -51,8 +60,16 @@ class CartItemsController extends Controller
         }
     }
 
-    public function update($id)
+    public function update(User $user, $id)
     {
+        if (Auth::user()->id != $user->id) {
+            return response()->json([
+                'status' => 'unauthorized',
+                'message' => '권환이 없습니다.',
+                'error' => '401'
+            ], 401);
+        }
+
         $this->validate(request(), [
             'quantity' => ['required', 'numeric'],
         ]);
@@ -67,14 +84,22 @@ class CartItemsController extends Controller
         ], 200);
     }
 
-    public function destroy($id)
+    public function destroy(User $user, $id)
     {
-        $cartItem = Auth::user()->cart->items()->findOrFail($id);
+        if (Auth::user()->id != $user->id) {
+            return response()->json([
+                'status' => 'unauthorized',
+                'message' => '권환이 없습니다.',
+                'error' => '401'
+            ], 401);
+        }
+
+        $cartItem = $user->cart->items()->findOrFail($id);
         $cartItem->delete();
 
         return response()->json([
             'status' => 'success',
-            'message' => '카트 아이템 수량변경 성공했어요.',
+            'message' => '카트 아이템 삭제 성공했어요.',
         ], 204);
     }
 }
